Refactor dashboard niche analytics into classic presenter
