+++ tests/ui/dashboardModel.test.js
@@ -94,4 +94,75 @@ test('buildDashboardViewModel produces derived dashboard sections', t => {
 
   const timeSummary = viewModel.dailyStats.time.summary;
   assert.ok(timeSummary.includes('invested'));
+
+  assert.ok(viewModel.niche, 'expected niche view model');
+  assert.equal(viewModel.niche.watchlistCount, 0);
+  assert.ok(Array.isArray(viewModel.niche.board.entries));
+  assert.ok(viewModel.niche.board.entries.every(entry => entry.assetCount === 0));
+  assert.equal(typeof viewModel.niche.highlights.hot.title, 'string');
+  assert.notEqual(viewModel.niche.highlights.hot.title, '');
+});
+
+test('buildDashboardViewModel includes niche analytics data', t => {
+  const state = buildDefaultState();
+  state.day = 5;
+  state.money = 2000;
+  state.timeLeft = 4;
+
+  state.niches = {
+    popularity: {
+      techInnovators: { score: 82, previousScore: 70 },
+      healthWellness: { score: 38, previousScore: 44 }
+    },
+    watchlist: ['healthWellness'],
+    lastRollDay: state.day
+  };
+
+  const stubAsset = {
+    id: 'niche-asset-test',
+    name: 'Niche Analytics Rig',
+    singular: 'Analytics Rig',
+    maintenance: { cost: 10 }
+  };
+
+  registry.assets.push(stubAsset);
+  t.after(() => {
+    registry.assets.pop();
+  });
+
+  state.assets[stubAsset.id] = {
+    instances: [
+      {
+        id: 'instance-1',
+        status: 'active',
+        nicheId: 'techInnovators',
+        lastIncome: 150,
+        lastIncomeBreakdown: {
+          total: 150,
+          entries: [
+            { type: 'niche', amount: 45 }
+          ]
+        }
+      }
+    ]
+  };
+
+  const summary = createSummary();
+  const viewModel = buildDashboardViewModel(state, summary);
+
+  const nicheModel = viewModel.niche;
+  assert.ok(nicheModel, 'expected niche model to exist');
+  assert.equal(nicheModel.watchlistCount, 1);
+
+  const entries = nicheModel.board.entries;
+  assert.ok(entries.length > 0, 'expected niche entries');
+  const techEntry = entries.find(entry => entry.id === 'techInnovators');
+  assert.ok(techEntry, 'expected tech innovators entry');
+  assert.equal(techEntry.assetCount, 1);
+  assert.equal(techEntry.trendImpact, 45);
+  assert.equal(techEntry.netEarnings, 150);
+  assert.equal(techEntry.status, 'Heating Up');
+
+  const highlightTitle = nicheModel.highlights.hot.title;
+  assert.ok(highlightTitle.includes('Tech Innovators'));
 });
